fix: update project to support Python 3.13
