+++ nearai/shared/near/sign.py
@@ -6,7 +6,6 @@ from enum import Enum
 from typing import Any, List, Optional, Union
 
 import base58
-import ed25519
 import nacl.signing
 import requests
 
@@ -266,24 +265,6 @@ def validate_completion_signature(public_key: str, signature: str, payload: Comp
         return False
 
 
-def derive_new_extended_private_key(extended_private_key: str, addition: str) -> str:
-    private_key_base58 = extended_private_key.replace("ed25519:", "")
-
-    decoded = base58.b58decode(private_key_base58)
-    secret_key = decoded[:32]
-
-    combined = secret_key + addition.encode()
-    derived_secret_key = hashlib.sha256(combined).digest()[:32]  # 32 bytes
-
-    new_signing_key = ed25519.SigningKey(derived_secret_key)
-
-    new_secret_key_bytes = new_signing_key.to_bytes() + new_signing_key.get_verifying_key().to_bytes()
-
-    new_private_key_base58 = base58.b58encode(new_secret_key_bytes[:64]).decode()
-
-    return f"ed25519:{new_private_key_base58}"
-
-
 def create_inference_signature(private_key: str, payload: CompletionSignaturePayload) -> tuple[str, str]:
     """Creates a cryptographic signature for a given extended inference payload using a specified private key."""
     borsh_payload = BinarySerializer(dict(COMPLETION_PAYLOAD_SCHEMA)).serialize(payload)
@@ -308,17 +289,3 @@ def create_inference_signature(private_key: str, payload: CompletionSignaturePay
     full_public_key = ED_PREFIX + public_key_base58
 
     return signature, full_public_key
-
-
-def get_public_key(extended_private_key):
-    private_key_base58 = extended_private_key.replace("ed25519:", "")
-
-    decoded = base58.b58decode(private_key_base58)
-    secret_key = decoded[:32]
-
-    signing_key = ed25519.SigningKey(secret_key)
-    verifying_key = signing_key.get_verifying_key()
-
-    base58_public_key = base58.b58encode(verifying_key.to_bytes()).decode()
-
-    return f"ed25519:{base58_public_key}"
