+++ pyproject.toml
@@ -6,7 +6,7 @@ authors = [
     { name = "Support", email = "support@near.ai" }
 ]
 readme = "README.md"
-requires-python = ">=3.9,<3.12"
+requires-python = ">=3.9,<3.14"
 dependencies = [
     "backoff>=2.2.1,<3.0.0",
     "base58==2.1.1",
@@ -15,9 +15,8 @@ dependencies = [
     "cryptography>=43.0.0",
     "datasets>=2.20.0,<3.0.0",
     "ddtrace==2.21.0",
-    "ed25519>=1.5,<2.0",
     "fastapi>=0.111.0,<0.112.0",
-    "fire>=0.6.0,<0.7.0",
+    "fire>=0.7.0",
     "jinja2>=3.1.4,<4.0.0",
     "litellm>=1.60.5,<2.0.0",
     "mypy-boto3>=1.34.147,<2.0.0",
@@ -52,11 +51,12 @@ dependencies = [
     "uvicorn>=0.32.0,<0.33.0",
     "tweepy>=4.14.0,<5.0.0",
     "rich>=13.7.0",
-    "py-near>=1.1.50,<2.0.0",
+    "py_near_primitives>= 0.2.3",
     "loguru>=0.7.2,<0.8.0",
     "ftfy>=6.3.1",
     "types-python-dateutil>=2.9.0.20241206",
     "wheel>=0.45.1",
+    "py-near @ git+https://github.com/pvolnov/py-near.git "
 ]
 
 [project.optional-dependencies]
@@ -67,6 +67,7 @@ explatform = [
     "shortuuid>=1.0.0,<2.0.0"
 ]
 hub = [
+    "ed25519>=1.5,<2.0",
     "fastapi-cli>=0.0.4,<1.0.0",
     "python-dotenv>=1.0.1,<2.0.0",
     "sqlmodel>=0.0.21,<1.0.0",
@@ -117,6 +118,9 @@ test = [
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+[tool.hatch.metadata]
+allow-direct-references = true
 
 [tool.setuptools]
 packages = ["nearai"]
@@ -235,7 +239,6 @@ module = [
     "torchao.*",
     "fireworks.*",
     "apscheduler.*",
-    "ed25519.*",
     "numpy.*",
     "nacl.*",
     "nacl.signing.*",
